商品发布

+++ toucan-shopping-cloud/apps/toucan-shopping-seller-web/src/main/java/com/toucan/shopping/cloud/apps/seller/web/vo/selectPage/GridResult.java
@@ -6,9 +6,21 @@ import java.util.List;
 
 @Data
 public class GridResult {
-    private Long pageSize;
+    /**
+     * 每页显示数量
+     */
+    private Integer pageSize;
+    /**
+     * 页码
+     */
     private int pageNumber;
+    /**
+     * 记录总数
+     */
     private Long totalRow;
+    /**
+     * 总页数
+     */
     private Long totalPage;
 
     private List<Object> list;
